Drop reversion of fix for upstream 2706 (#2323656)
We revert the fix for
https://gitlab.gnome.org/GNOME/mutter/-/issues/2706 /
https://bugzilla.redhat.com/show_bug.cgi?id=2323656 because it
causes https://gitlab.gnome.org/GNOME/mutter/-/issues/3068 /
https://bugzilla.redhat.com/show_bug.cgi?id=2239128 , which was
accepted as a release blocker for Fedora 39. But that bug is
specific to the installer (and, we think, to the installer on
Workstation live, at least there's no indication anyone reproduced
it in the dedicated installer environment). Now Fedora 41 is
released, we can drop the reversion so the other bug is fixed for
installed systems.

This will cause the installer bug to appear in the unofficial
live respins, which is unfortunate, but we judge that it's still
better to fix the other bug for installed systems.